pyproject.toml
[project]
name = "fragrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse==1.4.0",
    "chardet==5.2.0", 
    "faiss-cpu==1.9.0.post1",
    "gradio==5.12.0",
    "langchain==0.3.15",
    "langchain-community==0.3.15",
    "langchain-core==0.3.31",
    "langchain-huggingface==0.1.2",
    "openai==1.60.0",
    "nltk==3.9.1",
    "pypdf==5.1.0",
    "python-dotenv==1.0.1",
    "python-frontmatter==1.1.0",
]
